untested cleanup code
