+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonom/SampleAutonom.java
@@ -1,0 +1,65 @@
+package org.firstinspires.ftc.teamcode.autonom;
+
+import com.acmerobotics.dashboard.FtcDashboard;
+import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
+import com.acmerobotics.roadrunner.geometry.Pose2d;
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.hardware.DcMotorEx;
+import com.qualcomm.robotcore.hardware.Servo;
+
+import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
+import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
+import org.firstinspires.ftc.teamcode.teleop.DcMotorServo;
+import org.openftc.easyopencv.OpenCvCamera;
+import org.openftc.easyopencv.OpenCvCameraFactory;
+import org.openftc.easyopencv.OpenCvCameraRotation;
+import org.openftc.easyopencv.OpenCvWebcam;
+
+public abstract class SampleAutonom extends LinearOpMode {
+
+    DcMotorEx carusel;
+    DcMotorServo lift;
+    Servo cupa;
+    SampleMecanumDrive drive;
+
+    OpenCvWebcam webcam;
+    DetectionPipeline pipeline = new DetectionPipeline();
+
+    Pose2d startPose = new Pose2d(-35, -61.7, Math.toRadians(90));
+
+    public void initHardware(){
+        telemetry = new MultipleTelemetry(telemetry, FtcDashboard.getInstance().getTelemetry());
+
+        drive = new SampleMecanumDrive(hardwareMap);
+        drive.setPoseEstimate(startPose);
+
+        lift = new DcMotorServo(hardwareMap,"lift",13.79f,28);
+
+        carusel = hardwareMap.get(DcMotorEx.class, "carusel");
+        cupa = hardwareMap.get(Servo.class, "cupa");
+
+        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier(
+                "cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
+        webcam = OpenCvCameraFactory.getInstance().createWebcam(
+                hardwareMap.get(WebcamName.class, "Webcam 1"), cameraMonitorViewId);
+        webcam.setPipeline(pipeline);
+        webcam.setMillisecondsPermissionTimeout(2500); // Timeout for obtaining permission is configurable. Set before opening.
+        webcam.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener()
+        {
+            @Override
+            public void onOpened()
+            {
+                webcam.startStreaming(320, 240, OpenCvCameraRotation.UPRIGHT);
+            }
+
+            @Override
+            public void onError(int errorCode)
+            {
+
+            }
+        });
+    }
+
+    @Override
+    public abstract void runOpMode() throws InterruptedException;
+}
